Reorganize for VZCode use
