+++ package.json
@@ -6,12 +6,15 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "edit": "cd viz; vzcode"
   },
   "devDependencies": {
-    "vite": "^4.4.5"
+    "vite": "^4.4.9",
+    "vzcode": "^0.11.0"
   },
   "dependencies": {
+    "d3-color": "^3.1.0",
     "d3-ease": "^3.0.1",
     "d3-selection": "^3.0.0",
     "d3-transition": "^3.0.1"
